pyproject.toml
[project]
name = "omi-sync"
version = "0.1.0"
description = "Sync Omi conversations to Obsidian vault"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.1.0",
    "python-dateutil>=2.8.0",
    "pytz>=2024.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-httpx>=0.30.0",
    "freezegun>=1.4.0",
]

[project.scripts]
omi-sync = "omi_sync.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/omi_sync"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
